refactor: split community endpoint helpers
- Moves endpoint URL validation and OpenAI route construction into a focused module.
- Moves bounded image and video response handling into a focused media module.
- Updates consumers to import helpers directly without compatibility re-exports.
