+++ enter.pollinations.ai/src/routes/community-endpoints.ts
@@ -1,3 +1,4 @@
+import { validateCommunityEndpointUrl } from "@shared/community-endpoint-urls.ts";
 import {
     type CommunityEndpointVisibility,
     communityModelId,
@@ -10,7 +11,6 @@ import {
     parseListingPayload,
     pendingCommunityEndpointChangeIsReady,
     resolveEffectiveProxyListing,
-    validateCommunityEndpointUrl,
 } from "@shared/community-endpoints.ts";
 import * as schema from "@shared/db/better-auth.ts";
 import { validator } from "@shared/middleware/validator.ts";
