+++ gen.pollinations.ai/src/audio/communityEndpoint.ts
@@ -1,7 +1,7 @@
+import { communityAudioTranscriptionsUrl } from "@shared/community-endpoint-urls.ts";
 import {
     COMMUNITY_ENDPOINT_TIMEOUT_MS,
     type CommunityEndpointRuntime,
-    communityAudioTranscriptionsUrl,
     communityTranscriptionSeconds,
     normalizeCommunityEndpointBearerToken,
 } from "@shared/community-endpoints.ts";
